% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimi_stics.R
\name{opti}
\alias{opti}
\title{Optimization function}
\usage{
opti(USM_path, Plant, weight, obs_name, Parameters, ...)
}
\arguments{
\item{USM_path}{Path to the USM directory for evaluation.}

\item{Plant}{A vector for the plant (\emph{i.e.} Principal or associated) for which the parameters
will be set (only for plant or technical parameters in mixed crop simulations)
Set to \code{NULL} if using STICS in sole crop}

\item{weight}{The weight used for each variable (see details)}

\item{obs_name}{A \code{data.frame} of observation file name(s) of the form
\verb{data.frame(Principal= c(obs1.obs), Dominated= c(obs2.obs)} for mixed crops (simply
remove the \code{Dominated} column for sole crops.}

\item{Parameters}{A data.frame with parameter name, starting (optional), min, max values, and data type (optional). See details and example.}

\item{...}{Further arguments to pass to the optimization function, i.e. \code{\link[dfoptim:nmkb]{dfoptim::nmkb()}} or \code{\link[stats:optimize]{stats::optimize()}}}
}
\value{
A named vector of the optimized values of the parameters.
}
\description{
Optimization function
}
\details{
The function uses \code{\link[stats:optimize]{stats::optimize()}} for univariate optimization, and the \pkg{dfoptim} package functions for multivariate.
Currently only the Nelder-Mead algorithm is implemented from \pkg{dfoptim}.
The \code{Parameters} argument should be formated as a a data.frame (see example).
The start values should exclude the min and max values (they are exclusive bounds).
If the start is \code{NULL}, then the mean value between the min and max values is taken. The data type is optional
and only takes double (numeric) or integer. If a parameter is an integer, then a rounding is applied (very crude but functional).
If weight is not provided by the user, the selection criteria is computed using the equation
5 from Wallach et al. (2011). If they are provided, the equation 6 is used instead.
}
\references{
Wallach, D., Buis, S., Lecharpentier, P., Bourges, J., Clastre, P., Launay, M., … Justes, E. (2011).
A package of parameter estimation methods and implementation for the STICS crop-soil model. Environmental Modelling & Software, 26(4), 386–394. doi:10.1016/j.envsoft.2010.09.004
}
\seealso{
\code{\link[dfoptim:nmkb]{dfoptim::nmkb()}}
}
